+++ src/components/VideoProvider/index.tsx
@@ -1,8 +1,7 @@
 /**
  * #ANGLES_EDIT
  *
- * Add HuddleProvider inside the VideoContext.
- * Add WidgetProvider inside the VideoContext.
+ * Add RoomStateProvider.
  * `localTracks` context property can now include LocalDataTracks.
  */
 
@@ -18,12 +17,6 @@ import useLocalTracks from './useLocalTracks/useLocalTracks';
 import useRoom from './useRoom/useRoom';
 
 import { RoomStateProvider } from '../../withComponents/RoomState/RoomStateProvider';
-import { HuddleProvider } from '../../withComponents/HuddleProvider/HuddleProvider';
-import { WidgetProvider } from '../../withComponents/WidgetProvider/WidgetProvider';
-import { RoomMetaProvider } from '../../withComponents/RoomMetaProvider/RoomMetaProvider';
-import huddles from '../../withComponents/HuddleProvider/huddleReducer';
-import widgets from '../../withComponents/WidgetProvider/widgetReducer';
-import properties from '../../withComponents/RoomMetaProvider/roomMetaReducer';
 
 /*
  *  The hooks used by the VideoProvider component are different than the hooks found in the 'hooks/' directory. The hooks
@@ -77,14 +70,8 @@ export function VideoProvider({ options, children, onError = () => {}, onDisconn
         connect,
       }}
     >
-      <RoomStateProvider reducers={{ huddles, widgets, properties }}>
-        <RoomMetaProvider>
-          <HuddleProvider>
-            <WidgetProvider>
-              <SelectedParticipantProvider room={room}>{children}</SelectedParticipantProvider>
-            </WidgetProvider>
-          </HuddleProvider>
-        </RoomMetaProvider>
+      <RoomStateProvider>
+        <SelectedParticipantProvider room={room}>{children}</SelectedParticipantProvider>
       </RoomStateProvider>
     </VideoContext.Provider>
   );